requirements.txt
colour == 0.1.5
matplotlib == 3.4.2
numpy == 1.21.1
pandas == 1.3.1
Pillow == 8.3.1
scipy == 1.7.1
